Make the SVE kernels faster, one measured change at a time
The predicted SVE2 deficit was mostly two things, in roughly equal parts.
Disassembling md5_sve2_s1 against md5_neon_s1: identical vector work, 260
adds each, but 154 address instructions against NEON's 83, and 64 movprfx
NEON does not need.

Fixed, and what each was worth:

  md5    SVE 1.37x -> 1.68x    SVE2 0.76x -> 1.03x
  sha1   SVE 1.33x -> 1.33x    SVE2 0.80x -> 0.92x
  sha512 SVE 1.32x -> 1.33x    SVE2 0.89x -> 0.91x

**Round constants through a base register.** svdup of a table element
compiles to ld1rw, which loads and broadcasts in one instruction and takes
an immediate offset ample for a 256-byte table -- but GCC re-materialises
the table address for all 64 constants, so each costs adrp + add + ld1rw.
An empty asm makes the pointer opaque and the address is formed once.

The live range is the whole point. Declared at function scope this is 30%
*worse* than doing nothing: holding a register across the entire nest
costs more in spills than the addressing saves. Declared inside the block
loop, where the constants are actually read, it is the largest single win
here. Both were measured; the first was my first guess.

**XAR for the rotate.** Xor-then-rotate-right in one instruction, and with
a zero operand it is a plain rotate, so a left rotate by n costs one
instruction rather than shift-left plus shift-right-and-insert.

**EOR3 and XAR where a three-way XOR meets a rotate.** SHA-1's schedule
expansion is exactly that shape and drops from five instructions to two.
The templates now ask for XOR3 and XORROT by name and compose them from
two-input operations when the ISA has neither, in the statement form they
already used, so no fixed-width ISA moves.

**No EOR3 in SHA-512, deliberately.** Its four sigmas are three-way XORs
and this should have been free money; measured, it costs 3%. The operands
are rotations of one value that stay live across all four sigmas, and EOR3
is destructive.

Two things tried and rejected on measurement: page-aligning the constant
table so adrp alone would reach it, which GCC ignores, and function-scope
pinning as above.

The remaining SVE2 deficit on SHA-1 and SHA-512 is the schedule window.
Vector-length agnostic code cannot use compile-time byte offsets, so the
window is addressed with [base, #imm, mul vl] -- which GCC does use, 91
times -- but that immediate reaches only eight vectors, and the window is
sixteen. That one is structural.

x86 codegen is still byte-identical, 313432 instructions. 72 SVE and SVE2
kernel/vector-length combinations still match the scalar reference.

+++ src/kernels/cpu/sve_sha1_hooks.h
@@ -41,7 +41,7 @@
 
 /* The schedule window: a scalar buffer, strided by the run-time lane count. */
 #define S1K_WDECL           uint32_t wbuf[S1K_STREAMS * 16 * VB_MAX_LANES];
-#define S1K_WAT(k, i)       (wbuf + (((k) * 16 + (i)) * (size_t) S1K_LANES))
+#define S1K_WAT(k, i)       (wbuf + ((size_t) ((k) * 16 + (i)) * (size_t) S1K_LANES))
 #define S1K_WGET(k, i)      S1K_LOAD(S1K_WAT(k, i))
 #define S1K_WSET(k, i, v)   S1K_STORE(S1K_WAT(k, i), (v))
